1.新增判斷用戶是否登入時勾選'記住用戶名'功能 有勾選→登出時不刪除session 無勾選→登出時刪除session

+++ member_system/delete.php
@@ -1,7 +1,11 @@
 <?php
 include_once "../PDO.php";
-// 刪除資料庫內整筆會員資料
+// 判斷輸入帳號與密碼都正確就刪除資料庫內整筆會員資料
+if($_POST['acc']==$_SESSION['acc'] && $_POST['pass']==$_SESSION['pass']){
 $pdo=$pdo->exec("DELETE FROM `member_login` WHERE `member_login`.`member_id`='{$_SESSION['id']}'");
 session_destroy();
 header("location:login.php");
+}else{
+  header("location:member_data.php");
+}
 ?>
